pyproject.toml
[tool.poetry]
name = "rewrite-sample"
version = "0.1.0"
description = ""
authors = ["Sherub Thakur <sherub.thakur@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
blinker = "1.4"
datadog = "0.38.0"
flask = "1.1.2"
flask-sqlalchemy = "2.4.4"
psycopg2-binary = "2.8.5"
requests = "2.24.0"
sentry-sdk = "0.16.0"
simple-worker = "0.2.21"
sqlalchemy = "1.3.18"
srv-hijacker = "0.0.9"
stringcase = "1.2.0"
structlog = "20.1.0"
voluptuous = "0.11.7"
pyshuttlis = "0.1.50"
ddtrace = "0.36.0"
bookmarker = "0.0.1"
dnspython = "2.1.0"
flask-keepincheck = "0.1"
sqlalchemy-uuidstr = "^0.0.2"

[tool.poetry.dev-dependencies]
colorama = "0.4.3"
pytest = "5.4.3"
pytest-cov = "2.10.0"
pytest-flask = "1.0.0"
pytest-pep8 = "1.0.6"
pytest-runner = "5.2"
responses = "0.10.15"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
